+++ src/SlmQueue/Queue/ProvidesQueue.php
@@ -2,31 +2,10 @@
 
 namespace SlmQueue\Queue;
 
+/**
+ * @deprecated Use QueueAwareTrait instead
+ */
 trait ProvidesQueue
 {
-    /**
-     * @var QueueInterface
-     */
-    protected $queue;
-
-    /**
-     * Retrieve the queue
-     *
-     * @return QueueInterface
-     */
-    public function getQueue()
-    {
-        return $this->queue;
-    }
-
-    /**
-     * Inject a Queue instance
-     *
-     * @param  QueueInterface $queue
-     * @return void
-     */
-    public function setQueue(QueueInterface $queue)
-    {
-        $this->queue = $queue;
-    }
+    use QueueAwareTrait;
 }