+++ CHANGELOG.md
@@ -6,6 +6,8 @@ your `composer.json` file!
 to call parent constructor.
 - BC: keys for configuring queues was previously "queues", it is now "queue_manager". The key "queues" is still used
 but it's now for specifying options for a specific queue.
+- BC: remove Version class
+- BC: ProvidesQueue trait has been renamed to QueueAwareTrait, to provide PSR compliance
 - Job metadata is now serialized
 - You can make your jobs implement the interface 'SlmQueue\Queue\QueueAwareInterface'. Therefore, you will have
 access to the queue in the `execute` method of the job.